+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changes in this release:
 - Add support for skip in reference
 - Fix memory leak caused by pytest LogCaptureHandler.  Address it by clearing the old in-memory logs before each compile.
 - Fix incompatibility with versions of inmanta-core that have support to push environment settings to the server.
+- Reset the compiler state between two consecutive compiles
 
 # v 3.2.0 (2025-04-09)
 Changes in this release: